if a :slug lookup fails redirect to search
this method doesn't work perfect as umlauts are dropped during
slugging so finding them is a problem and searching for a slugged
entry is also silly.

Try farhang.im/Wörterbuch and you won't find it because it is
under farhang/im/woerterbuch. However, if you try farhang.im/woerterb
it will not find Wörterbuch because oe != ö. Anyway, edge cases.